Move controller init to confirmations folder
